+++ mayan/apps/file_metadata_msg/tests/test_drivers.py
@@ -2,23 +2,23 @@ from mayan.apps.common.tests.literals import (
     TEST_ARCHIVE_MSG_STRANGE_DATE_PATH
 )
 from mayan.apps.documents.tests.base import GenericDocumentTestCase
+from mayan.apps.file_metadata.tests.mixins import DocumentFileMetadataTestMixin
 
-from ..drivers.extract_msg import ExtractMSGToolDriver
+from ..drivers import FileMetadataDriverExtractMSGTool
 
 from .literals import (
     TEST_MSG_FILE_METADATA_DOTTED_NAME_SUBJECT,
     TEST_MSG_FILE_METADATA_DOTTED_NAME_TO,
     TEST_MSG_FILE_METADATA_VALUE_SUBJECT, TEST_MSG_FILE_METADATA_VALUE_TO
 )
-from .mixins import DocumentFileMetadataTestMixin
 
 
-class ExtractMSGDriverTestCase(
+class FileMetadataDriverExtractMSGTestCase(
     DocumentFileMetadataTestMixin, GenericDocumentTestCase
 ):
     _test_document_file_metadata_driver_create_auto = True
     _test_document_file_metadata_driver_enable_auto = True
-    _test_document_file_metadata_driver_path = ExtractMSGToolDriver.dotted_path
+    _test_document_file_metadata_driver_path = FileMetadataDriverExtractMSGTool.dotted_path
     _test_document_filename = TEST_ARCHIVE_MSG_STRANGE_DATE_PATH
 
     def test_driver_entries(self):